pyproject.toml
[project]
name = "aceteam-nodes"
version = "0.1.0"
description = "AceTeam workflow nodes for local execution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "aceteam-workflow-engine>=1.1.0rc1",
    "litellm>=1.40.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.11.0",
    "overrides>=7.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "pyright>=1.1",
]

[project.scripts]
aceteam-nodes = "aceteam_nodes.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aceteam_nodes"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
